rename from autoneal to auto spatial advisory
