+++ api/app/auto_spatial_advisory/polygonize.py
