+++ api/app/auto_spatial_advisory/process_hfi.py
@@ -12,8 +12,8 @@ from app import config
 from app.db.models.advisory import ClassifiedHfi, RunTypeEnum
 from app.db.database import get_async_write_session_scope
 from app.db.crud.fba_advisory import save_hfi
-from app.autoneal.classify_hfi import classify_hfi
-from app.autoneal.polygonize import polygonize
+from app.auto_spatial_advisory.classify_hfi import classify_hfi
+from app.auto_spatial_advisory.polygonize import polygonize
 from app.geospatial import NAD83_BC_ALBERS
