+++ api/app/routers/fba.py
@@ -11,7 +11,7 @@ from app.auth import authentication_required, audit
 from app.db.models.advisory import RunTypeEnum
 from app.schemas.fba import FireCenterListResponse, FireZoneAreaListResponse, FireZoneArea
 from app.wildfire_one.wfwx_api import (get_auth_header, get_fire_centers)
-from app.autoneal.process_hfi import process_hfi, RunType
+from app.auto_spatial_advisory.process_hfi import process_hfi, RunType
 
 logger = logging.getLogger(__name__)
